@@ -12,64 +12,59 @@
     "dist",
     "types"
   ],
-  "main": "dist/index.js",
   "keywords": [
     "immutable",
     "frozen",
     "functional",
     "declarative"
   ],
-  "scripts": {
-    "dtslint": "dtslint --localTs node_modules/typescript/lib types",
-    "test": "gulp && yarn -s dtslint",
-    "test:watch": "gulp watch",
-    "prepublish": "gulp prepublish",
-    "perf-server": "`npm bin`/webpack-dev-server --config perf/webpack.config.js --hot"
-  },
+  "source": "lib/index.js",
+  "main": "dist/index.js",
+  "module": "dist/index.module.js",
+  "types": "types",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/substantial/updeep/issues"
+  },
+  "scripts": {
+    "benchmark": "NODE_ENV=production node --require \"@babel/register\" perf",
+    "build": "microbundle",
+    "dev": "microbundle watch",
+    "dtslint": "dtslint --localTs node_modules/typescript/lib types",
+    "eslint": "eslint --ext .js,.jsx,.ts,.tsx",
+    "lint": "yarn -s eslint . && yarn -s dtslint",
+    "test": "mocha --recursive --require \"@babel/register\" test"
   },
   "dependencies": {
     "lodash": "^4.2.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.7.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
+    "@babel/preset-env": "^7.7.1",
+    "@babel/preset-typescript": "^7.9.0",
+    "@babel/register": "^7.7.0",
     "@types/chai": "^4.1.7",
     "@types/lodash": "^4.14.119",
     "@types/mocha": "^7.0.1",
-    "@types/node": "^13.1.0",
-    "babel-core": "^6.4.0",
-    "babel-loader": "^6.2.1",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-stage-2": "^6.3.13",
+    "@types/node": "^13.9.8",
+    "@typescript-eslint/eslint-plugin": "^2.26.0",
+    "@typescript-eslint/parser": "^2.26.0",
     "benchmark": "^2.1.4",
     "chai": "^4.2.0",
-    "dtslint": "^2.0.0",
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb-base": "^11.1.3",
-    "eslint-config-prettier": "^6.3.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-prettier": "^2.0.1",
-    "exports-loader": "^0.7.0",
-    "gulp": "^3.6.0",
-    "gulp-babel": "^7.0.1",
-    "gulp-eslint": "^3.0.1",
-    "gulp-mocha": "^2.0.0",
-    "karma": "^4.3.0",
-    "karma-babel-preprocessor": "^7.0.0",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.5",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "karma-webpack": "^2.0.13",
+    "chalk": "^4.0.0",
+    "dtslint": "^3.4.1",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.1",
+    "eslint-plugin-expect-type": "^0.0.4",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-prettier": "^3.1.2",
+    "husky": "^4.2.3",
+    "lint-staged": "^10.1.1",
+    "microbundle": "^0.11.0",
     "mocha": "^7.0.0",
-    "phantomjs-prebuilt": "^2.1.14",
-    "prettier": "^1.1.0",
-    "rimraf": "^3.0.0",
-    "typescript": "^3.6.3",
-    "webpack": "^1.10.5",
-    "webpack-dev-server": "^3.1.11",
-    "webpack-stream": "^2.1.0"
-  },
-  "types": "types"
+    "prettier": "^2.0.2",
+    "table": "^5.4.6",
+    "typescript": "^3.6.3"
+  }
 }
