@@ -29,4 +29,4 @@ u.updateIn = updateIn
 u.omitted = omitted
 u.withDefault = withDefault
 
-module.exports = u
+export default u
