@@ -13,7 +13,7 @@ function needsFreezing(object) {
 function recur(object) {
   Object.freeze(object)
 
-  Object.keys(object).forEach(key => {
+  Object.keys(object).forEach((key) => {
     const value = object[key]
     if (needsFreezing(value)) {
       recur(value)
