@@ -47,7 +47,7 @@ function resolveUpdates(updates, object) {
 function updateArray(updates, object) {
   const newArray = [...object]
 
-  Object.keys(updates).forEach(key => {
+  Object.keys(updates).forEach((key) => {
     newArray[key] = updates[key]
   })
 
@@ -91,13 +91,13 @@ function update(updates, object, ...args) {
 
   if (Array.isArray(defaultedObject)) {
     return updateArray(resolvedUpdates, defaultedObject).filter(
-      value => value !== innerOmitted
+      (value) => value !== innerOmitted
     )
   }
 
   return _omitBy(
     { ...defaultedObject, ...resolvedUpdates },
-    value => value === innerOmitted
+    (value) => value === innerOmitted
   )
 }
 
