@@ -7,7 +7,7 @@ const wildcard = '*'
 
 function reducePath(acc, key) {
   if (key === wildcard) {
-    return value =>
+    return (value) =>
       Object.prototype.hasOwnProperty.call(value, wildcard)
         ? // If we actually have wildcard as a property, update that
           update({ [wildcard]: acc }, value)
